Read power_current asynchronously
Reading /sys/class/power_supply/BAT0/power_now is somewhat expensive: it
requires some ACPI calls and takes several milliseconds (on my laptop
it's around 5). These calls are [cached by the kernel][1] for one
second, so reading several such files is as fast as reading one, but it
still means we're blocking the entire gnome-shell (the compositor is
there!) for a few milliseconds every time we refresh.

See:

    $ time read -r </sys/class/power_supply/BAT0/power_now; \
      time read -r </sys/class/power_supply/BAT0/power_now

    real    0m0,005s
    user    0m0,001s
    sys     0m0,001s

    real    0m0,000s
    user    0m0,000s
    sys     0m0,000s

It's probably not really a big deal, as at 60 Hz each frame takes ~17
ms, and we only block the shell for like 5, but it'd still be nice not
to do this, especially considering that with Wayland, we're not only
blocking the compositor but also inputs like mouse etc.

[1]: https://github.com/torvalds/linux/blob/8830280a69ddfdbba7fb24d79dce309817783c6a/drivers/acpi/battery.c#L62-L64